dev: make preview library
